Module theon::adjunct

Static groups of homogeneous data.

This module provides APIs for types that expose a rectangular and ordered set of homogeneous data. These types are typically array-like, but these APIs are not limited to arrays. Any type that provides ordered elements of the same type may be capable of supporting these APIs.

A type that implements these traits and operations is known as an adjunct. See the Adjunct trait for more.

Implementations for adjunct traits are provided for integrated foreign types when enabling geometry features. For example, implementations of Adjunct and other traits are provided for nalgebra types when the geometry-nalgebra feature is enabled.

Traits

Adjunct

Converged

Extend

Fold

FromItems

IntoItems

Map

Truncate

ZipMap