Struct theon::query::Line

pub struct Line<S> where
    S: EuclideanSpace,  {
    pub origin: S,
    pub direction: Unit<Vector<S>>,
}

Line.

Describes a line containing an origin point and a direction. Lines extend infinitely from their origin along their direction $\hat{u}$. Unlike Ray, the direction component of Line extends in both the positive and negative.

This representation is typically known as the vector form $P_0 + t\hat{u}$ where $t$ is some non-zero time of impact.

Fields
origin: S

The origin or contained point of the line.

direction: Unit<Vector<S>>

The unit direction(s) in which the line extends from its origin.

Implementations

impl<S> Line<S> where
S: EuclideanSpace,

pub fn x() -> Self where
S: FiniteDimensional,
S::N: Cmp<U0, Output = Greater>,

pub fn y() -> Self where
S: FiniteDimensional,
S::N: Cmp<U1, Output = Greater>,

pub fn z() -> Self where
S: FiniteDimensional,
S::N: Cmp<U2, Output = Greater>,

pub fn into_ray(self) -> Ray<S>

impl<S> Line<S> where
S: EuclideanSpace + FiniteDimensional<N = U2>,

pub fn slope(&self) -> Option<Scalar<S>>

pub fn x_intercept(&self) -> Option<Scalar<S>>

pub fn y_intercept(&self) -> Option<Scalar<S>>