Crate theon

Theon abstracts Euclidean spaces and geometric queries with support for popular linear algebra and spatial crates in the Rust ecosystem.

Modules

adjunct

Static groups of homogeneous data.

integration

Integration of external crates and foreign types.

ops

prelude

Re-exports commonly used types and traits.

query

Spatial queries.

space

Vector and affine spaces.

Traits

AsPosition

Immutable positional data.

AsPositionMut

Mutable positional data.

Functions

lerp

Linearly interpolates between two values.

Type Definitions

Position