Struct theon::query::Ray

pub struct Ray<S> where
    S: EuclideanSpace,  {
    pub origin: S,
    pub direction: Unit<Vector<S>>,
}

Ray or half-line.

Describes a decomposed line with an origin or initial point and a direction. Rays extend infinitely from their origin. The origin $P_0$ and the point $P_0 + \hat{u}$ (where $\hat{u}$ is the direction of the ray) form a half-line originating from $P_0$.

Fields
origin: S

The origin or initial point of the ray.

direction: Unit<Vector<S>>

The unit direction in which the ray extends from its origin.

Implementations

impl<S> Ray<S> where
S: EuclideanSpace,

pub fn into_line(self) -> Line<S>

pub fn reverse(self) -> Self

Reverses the direction of the ray.

Reversing a ray yields its opposite, with the same origin and the opposing half-line.